Fix main.ts circular references preventing Vite to do HMR
