+++ src/helpers/compile.ts
@@ -1,6 +1,6 @@
 import Compiler from "@/compiler/compiler";
 import { getAppSimpleMsgDlgId } from "./editor";
-import { vm } from "@/main";
+import { vm } from "@/helpers/appContext";
 import i18n from "@/i18n";
 import { useStore } from "@/store/store";
